Put back button into component
